requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
pycodestyle==2.10.0
pykerberos==1.2.4
pymongo==3.13.0
python-snappy==0.6.1
pytz==2022.6
sqlparse==0.2.4
